Facegenie/requirements.txt
asgiref==3.4.1
backcall==0.2.0
captcha==0.3
certifi==2021.10.8
charset-normalizer==2.0.7
checksum==1.0.4
colorama==0.4.4
crypto==1.4.1
debugpy==1.5.1
decorator==5.1.0
Django==3.2.9
djongo==1.3.6
entrypoints==0.3
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
Naked==0.1.31
nest-asyncio==1.5.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
pycryptodome==3.11.0
Pygments==2.10.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.2.4
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
pymongo[srv]
dnspython